bot_accounts e2e tests to TypeScript
