+++ e2e/cypress/tests/integration/bot_accounts/edit_bot_username_spec.ts
@@ -78,7 +78,7 @@ describe('Edit bot username', () => {
 
                 cy.wrap(newBotName);
             }
-        }).then((newBotName) => {
+        }).then((newBotName: any) => {
             // * Set alias for bot entry in bot list, this also checks that the bot entry exists
             cy.get('.backstage-list__item').contains('.backstage-list__item', newBotName).as('newbotEntry');
 
@@ -105,7 +105,7 @@ describe('Edit bot username', () => {
     const NAMING_WARNING_STANDARD = 'Usernames have to begin with a lowercase letter and be 3-22 characters long. You can use lowercase letters, numbers, periods, dashes, and underscores.';
     const NAMING_WARNING_ENDING_PERIOD = 'Bot usernames cannot have a period as the last character';
 
-    function tryUsername(name, warningMessage) {
+    function tryUsername(name: string, warningMessage?: string) {
         cy.get('#username').clear().type(name);
         cy.get('#saveBot').click();
