+++ e2e/cypress/tests/integration/bot_accounts/in_lists_1_spec.ts
@@ -10,6 +10,8 @@
 // Stage: @prod
 // Group: @bot_accounts
 
+import {UserProfile} from 'mattermost-redux/types/users';
+
 import {createBotPatch} from '../../support/api/bots';
 import {generateRandomUser} from '../../support/api/user';
 
@@ -40,8 +42,8 @@ describe('Bots in lists', () => {
 
             // # Create users
             createdUsers = await Promise.all([
-                client.createUser(generateRandomUser()),
-                client.createUser(generateRandomUser()),
+                client.createUser(generateRandomUser() as UserProfile),
+                client.createUser(generateRandomUser() as UserProfile),
             ]);
 
             await Promise.all([
@@ -52,8 +54,8 @@ describe('Bots in lists', () => {
                 cy.wrap(user).its('username');
 
                 // # Add to team and channel
-                await client.addToTeam(team.id, user.user_id ?? user.id);
-                await client.addToChannel(user.user_id ?? user.id, channel.id);
+                await client.addToTeam(team.id, user.user_id ? user.user_id : user.id);
+                await client.addToChannel(user.user_id ? user.user_id : user.id, channel.id);
             }));
         });
     });
