+++ e2e/cypress/tests/integration/bot_accounts/in_lists_2_spec.ts
@@ -9,6 +9,8 @@
 
 // Group: @bot_accounts
 
+import {UserProfile} from 'mattermost-redux/types/users';
+
 import {createBotPatch} from '../../support/api/bots';
 import {generateRandomUser} from '../../support/api/user';
 
@@ -47,20 +49,20 @@ describe('Bots in lists', () => {
 
             // # Create users
             const createdUsers = await Promise.all([
-                client.createUser(generateRandomUser()),
-                client.createUser(generateRandomUser()),
+                client.createUser(generateRandomUser() as UserProfile),
+                client.createUser(generateRandomUser() as UserProfile),
             ]);
 
             await Promise.all([
                 ...bots,
                 ...createdUsers,
-            ].map(async (user) => {
+            ].map(async (user: UserProfile & {user_id: string}) => {
                 // * Verify username exists
                 cy.wrap(user).its('username');
 
                 // # Add to team and channel
-                await client.addToTeam(team.id, user.user_id ?? user.id);
-                await client.addToChannel(user.user_id ?? user.id, channel.id);
+                await client.addToTeam(team.id, user.user_id ? user.user_id : user.id);
+                await client.addToChannel(user.user_id ? user.user_id : user.id, channel.id);
             }));
         });
     });
