+++ e2e/cypress/tests/integration/bot_accounts/promote_demote_spec.ts
